formulaire voiture/annonce

+++ classes/AuctionsClass.php
@@ -2,12 +2,12 @@
 
 class Auctions{
 
-   protected float $auctionprice;
+   protected int $auctionprice;
    protected string $auctiondate;
 //    id car ?
 // id user?
 
-protected function __construct(float $auctionprice, string $auctiondate)
+protected function __construct(int $auctionprice, string $auctiondate)
 {
     $this->auctionprice=$auctionprice;
     $this->auctiondate=$auctiondate;
